+++ quicksort.h
@@ -7,11 +7,11 @@ class QuickSort : public SortingAlgorithm
 {
 private:
     Type type = Type::QuickSort;
-    static void quickSort(std::vector<float>& data, int lewy, int prawy, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window);
+    static void quickSort(std::vector<int>& data, int lewy, int prawy, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window);
 
 public:
     QuickSort() = default;
-    void sort(std::vector<float>& data, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window);
+    void sort(std::vector<int>& data, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window);
     Type getType() { return type;}
 };
